refactor(nika-cli,nika-dap): descend the trust plane + drift compute to nika-dap — the 15k wall
Since #660 + #661 landed, nika-cli measures 15,692 prod LOC (cap
15,000 · scripts/ci/check-crate-size.sh) — red on main, and the queued
PRs (#662 evidence pack · #665 trace anchor) each need headroom the
crate does not have. Per the crate's own precedent (display, dap:
compute descends, render stays), the trust plane and the drift scan
move to their charter home — the trace-forensics plane — and the CLI
re-exports at the old paths so no call site changes.

What moved (all with their tests):

- crates/nika-cli/src/seal.rs → nika_dap::seal — the run-key custody
  (OS keychain first · 0600 fallback · env override), fingerprint,
  key init/trust/rotate, seal_event. nika-cli keeps `pub use` at
  `crate::seal::*`: verbs/key.rs, verbs/run/sink.rs read unchanged.
- the workflow author-binding half of that file → the NEW
  nika_dap::sign — sidecar_path, enrolled_pubboxes (sharing the seal
  module's ONE custody home), WorkflowSig, sign_workflow_with,
  check_workflow, verify_sidecar. verbs/sign.rs (clap + render) and
  the run --require-signature gate delegate through the same
  re-export, untouched.
- crates/nika-cli/src/verbs/check/drift.rs → nika_dap::drift — the
  NIKA-DRIFT-001 declared-vs-used scan (scan + UsedNames + the
  per-entry permit usage analysis). #661 had already documented why it
  could not live in nika-schema (that crate is at the cap); the check
  verb keeps the HINT block + --json rows via a pub(super) shim.
  (Named fallback in the mission: the trust plane alone shed 398 LOC,
  still red at 15,294 — drift's 540 closes it.)

Numbers (scripts/ci/check-crate-size.sh, prod LOC):

- nika-cli: 15,692 → 14,754 (under cap · ~250 headroom for #662)
- nika-dap: 2,470 → 3,408 (far under its own cap)

Deps: nika-dap gains nika-runtime (proof::preimage), nika-cap, url,
keyring, minisign; nika-cli drops keyring, minisign, nika-cap, url
(machete clean). main.rs untouched (1,316/1,500). public-api.txt:
nika-cli swaps the three `pub fn seal::key_*` lines for the two glob
re-exports (all other lines byte-kept — the zvariant NoneValue
blanket impls are the documented macOS-vs-ubuntu render skew CI
expects); nika-dap appends the seal/sign/drift surfaces (ubuntu
0.51.0 --omit auto-trait-impls form).

Gates: cargo test --workspace --lib (54 ok, 0 failed) · clippy
--all-targets -D warnings (0) · fmt --check · hygiene 36 green / 0 red
· cargo deny check ok · machete clean · crate-size green · all 8
ci-ratchets green at the pre-push gate.
