+++ crates/nika-cli/public-api.txt
@@ -15,9 +15,8 @@ pub use nika_cli::registry::Resolved
 pub use nika_cli::registry::is_registry_ref
 pub fn nika_cli::registry::resolve_blocking(arg: &str) -> core::result::Result<nika_registry_client::Resolved, nika_registry_client::RegistryError>
 pub mod nika_cli::seal
-pub fn nika_cli::seal::key_init(force: bool) -> core::result::Result<alloc::string::String, alloc::string::String>
-pub fn nika_cli::seal::key_rotate() -> core::result::Result<alloc::string::String, alloc::string::String>
-pub fn nika_cli::seal::key_trust() -> core::option::Option<(alloc::string::String, alloc::string::String)>
+pub use nika_cli::seal::<<nika_dap::seal::*>>
+pub use nika_cli::seal::<<nika_dap::sign::*>>
 pub mod nika_cli::verbs
 pub mod nika_cli::verbs::catalog
 pub fn nika_cli::verbs::catalog::run(json: bool, theme: nika_display::theme::Theme) -> nika_cli::verbs::VerbOutput
