+++ crates/nika-dap/public-api.txt
@@ -38,6 +38,9 @@ impl<T> typenum::type_operators::Same for nika_dap::chain::Verdict
 pub type nika_dap::chain::Verdict::Output = T
 pub const nika_dap::chain::CHAIN_GENESIS: &[u8]
 pub fn nika_dap::chain::walk(raw: &str) -> nika_dap::chain::Verdict
+pub mod nika_dap::drift
+pub const nika_dap::drift::DRIFT_CODE: &str
+pub fn nika_dap::drift::scan(wf: &nika_schema::raw::workflow::RawWorkflow) -> alloc::vec::Vec<alloc::string::String>
 pub mod nika_dap::otel
 pub fn nika_dap::otel::project(events: &[nika_event::event::Event], include_content: bool, chain: core::option::Option<&nika_dap::chain::Verdict>, exporter_version: &str) -> core::result::Result<alloc::string::String, alloc::string::String>
 pub mod nika_dap::recover
@@ -379,6 +382,43 @@ pub type nika_dap::retention::RetentionConfig::Output = T
 pub fn nika_dap::retention::collect(dir: &std::path::Path, cfg: &nika_dap::retention::RetentionConfig, now: std::time::SystemTime) -> core::option::Option<nika_dap::retention::GcReport>
 pub fn nika_dap::retention::newest_per_workflow(traces: &[nika_dap::store::TraceMeta]) -> alloc::collections::btree::set::BTreeSet<usize>
 pub fn nika_dap::retention::plan(traces: &[nika_dap::store::TraceMeta], cfg: &nika_dap::retention::RetentionConfig, now: std::time::SystemTime) -> alloc::vec::Vec<(usize, nika_dap::retention::Reason)>
+pub mod nika_dap::seal
+pub fn nika_dap::seal::fingerprint(pubkey_box: &str) -> alloc::string::String
+pub fn nika_dap::seal::key_init(force: bool) -> core::result::Result<alloc::string::String, alloc::string::String>
+pub fn nika_dap::seal::key_rotate() -> core::result::Result<alloc::string::String, alloc::string::String>
+pub fn nika_dap::seal::key_trust() -> core::option::Option<(alloc::string::String, alloc::string::String)>
+pub fn nika_dap::seal::load_signing_key() -> core::option::Option<(minisign::secret_key::SecretKey, alloc::string::String)>
+pub fn nika_dap::seal::seal_event(run_id: nika_types::id::EventId, at: nika_types::timestamp::Timestamp, head: &str, events: usize, workflow_hash: &str, engine: &str, sk: &minisign::secret_key::SecretKey, pk_box: &str) -> core::option::Option<nika_event::event::Event>
+pub mod nika_dap::sign
+pub enum nika_dap::sign::WorkflowSig
+pub nika_dap::sign::WorkflowSig::Invalid(alloc::string::String)
+pub nika_dap::sign::WorkflowSig::MissingSidecar
+pub nika_dap::sign::WorkflowSig::NoEnrolledKey
+pub nika_dap::sign::WorkflowSig::Valid(alloc::string::String)
+impl<D> owo_colors::OwoColorize for nika_dap::sign::WorkflowSig
+impl<T, U> core::convert::Into<U> for nika_dap::sign::WorkflowSig where U: core::convert::From<T>
+pub fn nika_dap::sign::WorkflowSig::into(self) -> U
+impl<T, U> core::convert::TryFrom<U> for nika_dap::sign::WorkflowSig where U: core::convert::Into<T>
+pub type nika_dap::sign::WorkflowSig::Error = core::convert::Infallible
+pub fn nika_dap::sign::WorkflowSig::try_from(value: U) -> core::result::Result<T, <T as core::convert::TryFrom<U>>::Error>
+impl<T, U> core::convert::TryInto<U> for nika_dap::sign::WorkflowSig where U: core::convert::TryFrom<T>
+pub type nika_dap::sign::WorkflowSig::Error = <U as core::convert::TryFrom<T>>::Error
+pub fn nika_dap::sign::WorkflowSig::try_into(self) -> core::result::Result<U, <U as core::convert::TryFrom<T>>::Error>
+impl<T> core::any::Any for nika_dap::sign::WorkflowSig where T: 'static + ?core::marker::Sized
+pub fn nika_dap::sign::WorkflowSig::type_id(&self) -> core::any::TypeId
+impl<T> core::borrow::Borrow<T> for nika_dap::sign::WorkflowSig where T: ?core::marker::Sized
+pub fn nika_dap::sign::WorkflowSig::borrow(&self) -> &T
+impl<T> core::borrow::BorrowMut<T> for nika_dap::sign::WorkflowSig where T: ?core::marker::Sized
+pub fn nika_dap::sign::WorkflowSig::borrow_mut(&mut self) -> &mut T
+impl<T> core::convert::From<T> for nika_dap::sign::WorkflowSig
+pub fn nika_dap::sign::WorkflowSig::from(t: T) -> T
+impl<T> nika_error::traits::AsAny for nika_dap::sign::WorkflowSig where T: 'static
+pub fn nika_dap::sign::WorkflowSig::as_any(&self) -> &(dyn core::any::Any + 'static)
+impl<T> typenum::type_operators::Same for nika_dap::sign::WorkflowSig
+pub type nika_dap::sign::WorkflowSig::Output = T
+pub fn nika_dap::sign::check_workflow(path: &std::path::Path) -> nika_dap::sign::WorkflowSig
+pub fn nika_dap::sign::sidecar_path(workflow: &std::path::Path) -> std::path::PathBuf
+pub fn nika_dap::sign::sign_workflow_with(path: &std::path::Path, sk: &minisign::secret_key::SecretKey, pk_box: &str) -> core::result::Result<alloc::string::String, alloc::string::String>
 pub mod nika_dap::source_id
 pub fn nika_dap::source_id::lf_normal_form(source: &str) -> alloc::borrow::Cow<'_, str>
 pub fn nika_dap::source_id::sha256_hex(bytes: &[u8]) -> alloc::string::String
