+++ crates/nika-dap/src/lib.rs
@@ -24,10 +24,13 @@
 //! → stopped/continue loop → disconnect.
 
 pub mod chain;
+pub mod drift;
 pub mod otel;
 pub mod recover;
 pub mod reproduce;
 pub mod retention;
+pub mod seal;
+pub mod sign;
 pub mod source_id;
 pub mod stats;
 pub mod store;
